requirements.txt
# Core
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.0
pyyaml>=6.0.1

# Validación
pydantic>=2.6.0

# CLI
typer[all]>=0.12.0

# Tests
pytest>=8.0.0
